membership-website-main/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.4.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
cassandra-driver==3.25.0
cffi==1.17.1
click==8.0.3
colorama==0.4.4
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Django==4.0
django-cassandra-engine==1.6.3
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-htmx==1.6.0
executing==2.2.0
fastjsonschema==2.21.1
geomet==0.2.1.post1
gunicorn==20.1.0
ipykernel==6.30.0
ipython==9.4.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.4.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
#pywin32==311
#pywinpty==2.0.15
pyzmq==27.0.1
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
rpds-py==0.26.0
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.7
sqlparse==0.4.2
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2021.5
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
